Handling arbitrary dimensions now
This enables the cubical complex module to deal with images of arbitrary
dimensions (provided the low-level interface supports them as well). The
main question is whether to preserve the *old* functionality or not---in
contrast to the Vietoris--Rips construction, the tensor information will
not be enough to determine dimensionality in most cases.