Supporting pre-computed distances
Restructured the Vietoris--Rips complex calculation so that distances
are not calculated *twice* any more. This prevents conceptual issues,
such as clients using `p = 1` while the complex is being created with
`p = 2`.

Moreover, the new interface addresses also issue #8 and permits input
data to be treated as representing matrices. This enables a client to
use *any* distance they prefer.