Added support for infinite features and thresholds
This substantially improves the versatility of the Vietoris--Rips complex:

- infinite features will now be returned
- potential support for thresholds is possible (not yet implemented)

On the other hand, some downstream analyses *break* now, so it is best
to turn this into a configurable option. I would argue that in case one
can compute the Vietoris--Rips complex, the single infinite point does
not really matter; this becomes much more interesting when threshold is
involved.